Update: AI behavior updated to improve level 4 alert stage ('Attack')
AI in level 4 alert stage can now detect whether the player is on the ground or on the wall to find the right way to intercept him. If the player is on the wall, the AI lives him and fires an EMI beam, which drains the battery that powers the grip system and causes the player to fall. Once on the ground, it moves towards the player to attack.